policy_holder_bot/tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
aiohttp==3.9.1
requests==2.31.0
python-telegram-bot==20.7
sqlalchemy==2.0.25
ic-py==0.0.10 